backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# pgvector support
pgvector==0.2.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.10.0
pydantic-settings==2.7.0
email-validator==2.2.0

# RAG & LLM
langchain==0.3.0
langchain-openai==0.2.0
langchain-google-genai==2.0.0
langchain-anthropic==0.3.0
langchain-community==0.3.0
langgraph==0.2.0
openai==1.50.0
tiktoken==0.7.0

# Agent Tools
duckduckgo-search==6.3.0

# Document Processing
pypdf==4.0.0
python-docx==1.1.0
markdown==3.5.2
pandas==2.2.0
openpyxl==3.1.2  # For .xlsx files
xlrd==2.0.1  # For .xls files

# OCR Support
pytesseract==0.3.10
Pillow==10.2.0
PyMuPDF==1.23.8
google-cloud-vision==3.5.0

# Utilities
aiofiles==23.2.1
httpx==0.26.0
nanoid==2.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
ruff==0.1.14
